refactor(syscalls): inline `SyscallInterface::get_application_parameters`
